Revert the change of neighbor_name property to optional
Due to BPSO-4421, the change from required to optional is currently not
allowed with existing resources.

I was testing this for the last PR and did not see any error in the
bpocore log, so thought it passed. Gordon is adding error logs when this
happens so it would not be missed.